Move list stuff into one file
